# [FIX] point_of_sale: improve get_price performance