@caburj 2022-04-29
@pimodoo Can we consider the original solution, where pricelist items are grouped by pricelist id? It wasn't wrong, and the benefit is that the algorithm is faster by a factor equal to the number of the pricelists. I know user's may defined just at most 10 pricelists, but what if they went for at least a 100? So it will 100x slower, and we will come back to this function, making same analysis and optimization. Let's save ourselves one trip.

cc @lse-odoo