@pimodoo 2022-04-29
yes, we can group item by pricelist