@lse-odoo 2022-04-29
@pimodoo perfect, did you notice anything else that you would like me to change ?

Note: I did double check on my standard instance and on the client's database locally